Add code for tuning hyperparameters using Keras Tuner
This commit adds code to tune the hyperparameters of a deep neural network using Keras Tuner. The code defines a function that returns a compiled model, which is then used by the tuner to search for the best model configuration. The results of the best model configuration are printed at the end of the script.